pyproject.toml
[project]
name = "vlk-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncio>=4.0.0",
    "discord-py>=2.6.4",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
]

# For npm-like scripts, use one of these methods:
# 1. Makefile: make dev, make prod
# 2. Python script: python run.py dev, python run.py prod
# 3. Direct: ENVIRONMENT=dev python main.py, ENVIRONMENT=prod python main.py
